home aboutセクションをページに記述し、コンポーネントを削除
