+++ src/public/css/app.css
@@ -1943,367 +1943,6 @@ import用ファイル
 /**
  * @keyframe を定義
  */
-@-webkit-keyframes iconSlide {
-0% {
-    transform: translateX(0);
-}
-10% {
-    transform: translateX(10%);
-}
-20% {
-    transform: translateX(20%);
-}
-30% {
-    transform: translateX(30%);
-}
-40% {
-    transform: translateX(40%);
-}
-50% {
-    transform: translateX(50%);
-}
-60% {
-    transform: translateX(40%);
-}
-70% {
-    transform: translateX(30%);
-}
-80% {
-    transform: translateX(20%);
-}
-90% {
-    transform: translateX(10%);
-}
-100% {
-    transform: translateX(0);
-}
-}
-@keyframes iconSlide {
-0% {
-    transform: translateX(0);
-}
-10% {
-    transform: translateX(10%);
-}
-20% {
-    transform: translateX(20%);
-}
-30% {
-    transform: translateX(30%);
-}
-40% {
-    transform: translateX(40%);
-}
-50% {
-    transform: translateX(50%);
-}
-60% {
-    transform: translateX(40%);
-}
-70% {
-    transform: translateX(30%);
-}
-80% {
-    transform: translateX(20%);
-}
-90% {
-    transform: translateX(10%);
-}
-100% {
-    transform: translateX(0);
-}
-}
-@-webkit-keyframes fadeoutDown {
-0% {
-    transform: translateY(0);
-}
-10% {
-    opacity: 0.9;
-    transform: translateY(10%);
-}
-20% {
-    opacity: 0.8;
-    transform: translateY(20%);
-}
-30% {
-    opacity: 0.7;
-    transform: translateY(30%);
-}
-40% {
-    opacity: 0.6;
-    transform: translateY(40%);
-}
-50% {
-    opacity: 0.5;
-    transform: translateY(50%);
-}
-60% {
-    opacity: 0.4;
-    transform: translateY(60%);
-}
-70% {
-    opacity: 0.3;
-    transform: translateY(70%);
-}
-80% {
-    opacity: 0.2;
-    transform: translateY(80%);
-}
-90% {
-    opacity: 0.1;
-    transform: translateY(90%);
-}
-100% {
-    opacity: 0;
-    transform: translateY(100%);
-}
-}
-@keyframes fadeoutDown {
-0% {
-    transform: translateY(0);
-}
-10% {
-    opacity: 0.9;
-    transform: translateY(10%);
-}
-20% {
-    opacity: 0.8;
-    transform: translateY(20%);
-}
-30% {
-    opacity: 0.7;
-    transform: translateY(30%);
-}
-40% {
-    opacity: 0.6;
-    transform: translateY(40%);
-}
-50% {
-    opacity: 0.5;
-    transform: translateY(50%);
-}
-60% {
-    opacity: 0.4;
-    transform: translateY(60%);
-}
-70% {
-    opacity: 0.3;
-    transform: translateY(70%);
-}
-80% {
-    opacity: 0.2;
-    transform: translateY(80%);
-}
-90% {
-    opacity: 0.1;
-    transform: translateY(90%);
-}
-100% {
-    opacity: 0;
-    transform: translateY(100%);
-}
-}
-@-webkit-keyframes fadeoutTop {
-0% {
-    opacity: 1;
-    transform: translateY(0);
-}
-100% {
-    opacity: 0;
-    transform: translateY(-30%);
-}
-}
-@keyframes fadeoutTop {
-0% {
-    opacity: 1;
-    transform: translateY(0);
-}
-100% {
-    opacity: 0;
-    transform: translateY(-30%);
-}
-}
-@-webkit-keyframes bound {
-0% {
-    transform: translate(0%, -100%);
-}
-30% {
-    transform: translate(0%, 0%) scale(1.1, 1);
-}
-40% {
-    transform: translate(0%, -40%);
-}
-55% {
-    transform: translate(0%, 0%) scale(1.1, 1);
-}
-60% {
-    transform: translate(0%, -15%);
-}
-70% {
-    transform: translate(0%, 0%) scale(1.1, 1);
-}
-75% {
-    transform: translate(0%, -5%);
-}
-95% {
-    transform: translate(0%, 0%) scale(1.1, 1);
-}
-100% {
-    transform: scale(1, 1);
-}
-}
-@keyframes bound {
-0% {
-    transform: translate(0%, -100%);
-}
-30% {
-    transform: translate(0%, 0%) scale(1.1, 1);
-}
-40% {
-    transform: translate(0%, -40%);
-}
-55% {
-    transform: translate(0%, 0%) scale(1.1, 1);
-}
-60% {
-    transform: translate(0%, -15%);
-}
-70% {
-    transform: translate(0%, 0%) scale(1.1, 1);
-}
-75% {
-    transform: translate(0%, -5%);
-}
-95% {
-    transform: translate(0%, 0%) scale(1.1, 1);
-}
-100% {
-    transform: scale(1, 1);
-}
-}
-@-webkit-keyframes rotation {
-0% {
-    transform: translateX(-50rem) rotate(0deg);
-}
-100% {
-    transform: translateX(0) rotate(1455deg);
-}
-}
-@keyframes rotation {
-0% {
-    transform: translateX(-50rem) rotate(0deg);
-}
-100% {
-    transform: translateX(0) rotate(1455deg);
-}
-}
-@-webkit-keyframes loading {
-0% {
-    height: 0;
-}
-50% {
-    height: 80%;
-}
-100% {
-    height: 0;
-}
-}
-@keyframes loading {
-0% {
-    height: 0;
-}
-50% {
-    height: 80%;
-}
-100% {
-    height: 0;
-}
-}
-/**
- * common mixin (サイト内でどこでも使う可能性がある)
- */
-/**
- * gradation
- */
-/**
- * element
- */
-/**
- * 装飾用 mixin
- */
-/**
- * page
- */
-/**
- * component
- */
-.text-box {
-  padding: 0 1rem 1rem 1rem;
-  background-color: rgba(255, 255, 255, 0.8);
-  box-shadow: 0px 6px 8px #e6e6e6;
-}
-@media screen and (min-width: 560px) {
-.text-box {
-    padding: 0 1.5rem 1.5rem 1.5rem;
-}
-}
-@media screen and (min-width: 1025px) {
-.text-box {
-    display: flex;
-    flex-flow: column nowrap;
-    justify-content: center;
-    align-items: center;
-    min-height: 25rem;
-}
-}
-.text-box__text {
-  margin-bottom: 2.5rem;
-}
-.text-box__button {
-  width: 100%;
-}
-@media screen and (min-width: 560px) {
-.text-box__button {
-    max-width: 25rem;
-    margin: 0 auto;
-}
-}@charset "UTF-8";
-/*
-import用ファイル
-グローバルに使用する変数やmixinをインポート
-*/
-/**
- * サイトで使用するマップを定義します
- */
-/**
- * ベースとなるフォントサイズ = 16px
- * modules.scssでremに変換して出力
- */
-/**
- * base pixel (8px)
- * rem だとフォントサイズに影響するため、pixelで定義
- */
-/**
- * google map size
- */
-/**
- * アスペクト比（基準を100%と想定）
- */
-/**
- * radius
- */
-/**
- * サイト内で扱うsass関数とキーフレームを定義するファイル。
- * _variables.scssで定義した変数を呼び出しやすい形に変換
- */
-/**
- * font size
- * @param 指定したいピクセル数を入力
- * @return 単位をremに変換し出力
- */
-/**
- * @keyframe を定義
- */
 @-webkit-keyframes iconSlide-data-v-10d75cdd {
 0% {
     transform: translateX(0);
@@ -3063,7 +2702,7 @@ import用ファイル
 }
 }
 @media screen and (min-width: 1025px) {
-.about:nth-child(even) .about__lead[data-v-63cd6604] {
+.about:nth-child(even) .about__box[data-v-63cd6604] {
     margin: 0 auto 0 0;
 }
 }
@@ -3094,23 +2733,54 @@ import用ファイル
     min-width: 33rem;
 }
 }
-.about__lead[data-v-63cd6604] {
+.about__box[data-v-63cd6604] {
   width: 90%;
   max-width: 50rem;
   margin: 0 auto;
   transform: translateY(-15%);
 }
 @media screen and (min-width: 560px) {
-.about__lead[data-v-63cd6604] {
+.about__box[data-v-63cd6604] {
     transform: translateY(-20%);
 }
 }
 @media screen and (min-width: 1025px) {
-.about__lead[data-v-63cd6604] {
+.about__box[data-v-63cd6604] {
     min-width: 30rem;
     width: 60%;
     margin: 0 0 0 auto;
     transform: translateY(-40%);
+}
+}
+.about__content[data-v-63cd6604] {
+  padding: 0 1rem 1rem 1rem;
+  background-color: rgba(255, 255, 255, 0.8);
+  box-shadow: 0px 6px 8px #e6e6e6;
+}
+@media screen and (min-width: 560px) {
+.about__content[data-v-63cd6604] {
+    padding: 0 1.5rem 1.5rem 1.5rem;
+}
+}
+@media screen and (min-width: 1025px) {
+.about__content[data-v-63cd6604] {
+    display: flex;
+    flex-flow: column nowrap;
+    justify-content: center;
+    align-items: center;
+    min-height: 25rem;
+}
+}
+.about__text[data-v-63cd6604] {
+  margin-bottom: 2.5rem;
+}
+.about__button[data-v-63cd6604] {
+  width: 100%;
+}
+@media screen and (min-width: 560px) {
+.about__button[data-v-63cd6604] {
+    max-width: 25rem;
+    margin: 0 auto;
 }
 }
 .support__text[data-v-63cd6604] {
