+++ src/public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=9ab87531d46fb475d031",
-    "/css/app.css": "/css/app.css?id=2bf0fc58684dc6be5e1c"
+    "/js/app.js": "/js/app.js?id=545663e12feedbf3b519",
+    "/css/app.css": "/css/app.css?id=6419ff1337c900029e49"
 }
